fix(70): fix reviews
