+++ lib/data/di/module/cart_data_module.dart
@@ -4,6 +4,8 @@ import 'package:mobile_ai_erp/data/network/apis/wishlist/wishlist_api.dart';
 import 'package:mobile_ai_erp/data/network/apis/coupon/coupon_api.dart';
 import 'package:mobile_ai_erp/data/repository/cart/cart_repository.dart';
 import 'package:mobile_ai_erp/data/repository/cart/cart_repository_impl.dart';
+import 'package:mobile_ai_erp/data/repository/wishlist/wishlist_repository.dart';
+import 'package:mobile_ai_erp/data/repository/wishlist/wishlist_repository_impl.dart';
 import 'package:mobile_ai_erp/data/repository/product/product_repository.dart';
 import 'package:mobile_ai_erp/data/repository/product/product_repository_impl.dart';
 import 'package:mobile_ai_erp/data/repository/coupon/coupon_repository.dart';
@@ -19,10 +21,13 @@ class CartDataModule {
   static void _setupRepository(GetIt getIt) {
     if (!getIt.isRegistered<CartRepository>()) {
       getIt.registerSingleton<CartRepository>(
-        CartRepositoryImpl(
-          cartApi: getIt<CartApi>(),
-          wishlistApi: getIt<WishlistApi>(),
-        ),
+        CartRepositoryImpl(cartApi: getIt<CartApi>()),
+      );
+    }
+
+    if (!getIt.isRegistered<WishlistRepository>()) {
+      getIt.registerSingleton<WishlistRepository>(
+        WishlistRepositoryImpl(wishlistApi: getIt<WishlistApi>()),
       );
     }
 
@@ -42,6 +47,10 @@ class CartDataModule {
       getIt.unregister<CartRepository>();
     }
 
+    if (getIt.isRegistered<WishlistRepository>()) {
+      getIt.unregister<WishlistRepository>();
+    }
+
     if (getIt.isRegistered<CouponRepository>()) {
       getIt.unregister<CouponRepository>();
     }
@@ -53,6 +62,7 @@ class CartDataModule {
 
   static bool isRegistered(GetIt getIt) {
     return getIt.isRegistered<CartRepository>() &&
+        getIt.isRegistered<WishlistRepository>() &&
         getIt.isRegistered<ProductRepository>() &&
         getIt.isRegistered<CouponRepository>();
   }
