+++ lib/data/repository/cart/cart_repository_impl.dart
@@ -1,21 +1,13 @@
 import 'package:mobile_ai_erp/data/network/apis/cart/cart_api.dart';
-import 'package:mobile_ai_erp/data/network/apis/wishlist/wishlist_api.dart';
 import 'package:mobile_ai_erp/data/repository/cart/cart_repository.dart';
 import 'package:mobile_ai_erp/domain/entity/cart/cart.dart';
 import 'package:mobile_ai_erp/domain/entity/cart/cart_calculation.dart';
 import 'package:mobile_ai_erp/domain/entity/cart/cart_item.dart';
-import 'package:mobile_ai_erp/domain/entity/cart/wishlist.dart';
-import 'package:mobile_ai_erp/domain/entity/cart/wishlist_item.dart';
 
 class CartRepositoryImpl implements CartRepository {
   final CartApi _cartApi;
-  final WishlistApi _wishlistApi;
 
-  CartRepositoryImpl({
-    required CartApi cartApi,
-    required WishlistApi wishlistApi,
-  }) : _cartApi = cartApi,
-       _wishlistApi = wishlistApi;
+  CartRepositoryImpl({required CartApi cartApi}) : _cartApi = cartApi;
 
   @override
   Future<Cart> getCart() async {
@@ -74,37 +66,6 @@ class CartRepositoryImpl implements CartRepository {
   Future<Cart> mergeCart({required List<Map<String, dynamic>> items}) async {
     final res = await _cartApi.mergeCart(items: items);
     return _mapCart(_unwrapData(res));
-  }
-
-  @override
-  Future<Wishlist> getWishlist() async {
-    final res = await _wishlistApi.getWishlist();
-    return _mapWishlist(_unwrapData(res));
-  }
-
-  @override
-  Future<Map<String, dynamic>> getWishlistSummary() async {
-    final res = await _wishlistApi.getWishlistSummary();
-    return _unwrapData(res);
-  }
-
-  @override
-  Future<Wishlist> addToWishlist({required String productId}) async {
-    final res = await _wishlistApi.addToWishlist(productId: productId);
-    return _mapWishlist(_unwrapData(res));
-  }
-
-  @override
-  Future<void> removeFromWishlist({required String itemId}) async {
-    await _wishlistApi.removeFromWishlist(itemId: itemId);
-  }
-
-  @override
-  Future<Wishlist> mergeWishlist({
-    required List<Map<String, dynamic>> items,
-  }) async {
-    final res = await _wishlistApi.mergeWishlist(items: items);
-    return _mapWishlist(_unwrapData(res));
   }
 
   Map<String, dynamic> _unwrapData(Map<String, dynamic> json) {
@@ -244,54 +205,6 @@ class CartRepositoryImpl implements CartRepository {
       isPriceChanged: json['isPriceChanged'] as bool? ?? false,
       isAvailable: json['isAvailable'] as bool? ?? false,
       stockWarning: json['stockWarning'] as bool? ?? false,
-    );
-  }
-
-  Wishlist _mapWishlist(Map<String, dynamic> json) {
-    final itemsJson = _asListOfMap(json['items']);
-    final wishlistId = (json['id'] ?? '').toString();
-
-    return Wishlist(
-      id: wishlistId,
-      tenantId: (json['tenantId'] ?? json['tenant_id'] ?? '').toString(),
-      customerId: (json['customerId'] ?? json['customer_id'] ?? '').toString(),
-      totalItems: (json['totalItems'] as num?)?.toInt() ?? 0,
-      items: itemsJson.map((e) => _mapWishlistItem(e, wishlistId)).toList(),
-      createdAt: DateTime.parse(
-        (json['createdAt'] ?? json['created_at']).toString(),
-      ),
-      updatedAt: DateTime.parse(
-        (json['updatedAt'] ?? json['updated_at']).toString(),
-      ),
-    );
-  }
-
-  WishlistItem _mapWishlistItem(Map<String, dynamic> json, String wishlistId) {
-    final attributesJson = _asListOfMap(json['attributes']);
-
-    return WishlistItem(
-      id: (json['id'] ?? '').toString(),
-      wishlistId: wishlistId,
-      productId: (json['productId'] ?? json['product_id'] ?? '').toString(),
-      variantId: json['variantId']?.toString(),
-      addedAt: DateTime.parse(json['addedAt'].toString()),
-      productName: (json['productName'] ?? '').toString(),
-      sku: (json['sku'] ?? '').toString(),
-      productType: (json['productType'] ?? '').toString(),
-      productStatus: (json['productStatus'] ?? '').toString(),
-      sellingPrice: (json['sellingPrice'] ?? '0').toString(),
-      originalPrice: json['originalPrice']?.toString(),
-      thumbnailUrl: json['thumbnailUrl']?.toString(),
-      variantSummary: json['variantSummary']?.toString(),
-      attributes: attributesJson
-          .map(
-            (attr) => WishlistItemAttribute(
-              label: (attr['label'] ?? '').toString(),
-              value: (attr['value'] ?? '').toString(),
-            ),
-          )
-          .toList(),
-      isAvailable: json['isAvailable'] as bool? ?? false,
     );
   }
 }
